Added step sound
Added simple step sound (borrowed together with the Sound() routine from Daniel C's Tiny-invaders).
Also enabled wrap-around, so levels do not need to end at the left or right border of the array. Liteally endless possibilites ;)
I still have to redo the graphics because when you stand directly in front of a wall it is not possible to see walls to the left or right which feels wrong.